fix: Send content type when uploading images
