+++ cli/internal/publish/images/testdata/content-type/png.png
